wildberries_service/requirements.txt
grpcio==1.71.0
grpcio-reflection==1.71.0
protobuf==5.29.4
python-dotenv==1.1.0